fix(desktop): use secure storage for WSClient JWT auth
WSClient was reading JWT from browser cookies, which may not exist due to
domain mismatch, httpOnly flags, or timing issues. This caused MCP tools
to fail with "Desktop app not connected" error.

Changes:
- Extract auth storage functions to new auth-storage.ts module
- WSClient now uses loadAuthSession() from secure storage (same as Socket.IO)
- Remove unused mainWindow dependency from WSClient
- Clean up unused imports in index.ts

The WSClient now uses the same token source as Socket.IO, ensuring
consistent authentication for both real-time and MCP tool connections.
